# Use deque for ~4x speedup when reading array columns