This PR increases performance when dealing with larger array columns by ~4x by using a `collections.deque` instead of `queue.Queue`.

Profiles are below. For reference, the clickhouse query took ~100ms to return the data and clickhouse_driver took 1300ms (queue.Queue) and 300ms (collections.deque) when processing the data to python objects.

I still think there may be room for improvement by lowering the amount of function calls, but maybe for another PR.

With queue.Queue

```bash
         1024218 function calls (1024203 primitive calls) in 1.382 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        3    0.397    0.132    1.318    0.439 arraycolumn.py:126(_read)
    30206    0.125    0.000    0.329    0.000 queue.py:121(put)
    60412    0.122    0.000    0.201    0.000 threading.py:335(notify)
    30206    0.112    0.000    0.310    0.000 queue.py:153(get)
    30209    0.048    0.000    0.075    0.000 queue.py:96(empty)
    30243    0.046    0.000    0.049    0.000 {method 'read' of 'clickhouse_driver.bufferedreader.BufferedReader' objects}
    30203    0.046    0.000    0.109    0.000 arraycolumn.py:45(size_unpack)
    60412    0.041    0.000    0.056    0.000 threading.py:243(__exit__)
    60412    0.041    0.000    0.079    0.000 threading.py:255(_is_owned)
    60412    0.039    0.000    0.067    0.000 threading.py:240(__enter__)
    60412    0.038    0.000    0.038    0.000 {method 'acquire' of '_thread.lock' objects}
    60415    0.035    0.000    0.050    0.000 queue.py:208(_qsize)
       38    0.034    0.001    0.034    0.001 {method 'read' of '_ssl._SSLSocket' objects}
    60412    0.029    0.000    0.029    0.000 {method '__enter__' of '_thread.lock' objects}
    30206    0.026    0.000    0.336    0.000 queue.py:192(get_nowait)
    52513    0.023    0.000    0.023    0.000 base.py:101(<genexpr>)
    30206    0.022    0.000    0.029    0.000 queue.py:212(_put)
    30216    0.020    0.000    0.020    0.000 {method 'unpack' of 'Struct' objects}
    30206    0.018    0.000    0.025    0.000 queue.py:216(_get)
```

With collections.deque
```bash
             299247 function calls (299232 primitive calls) in 0.430 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        3    0.180    0.060    0.370    0.123 arraycolumn.py:121(_read)
    30243    0.045    0.000    0.048    0.000 {method 'read' of 'clickhouse_driver.bufferedreader.BufferedReader' objects}
    30203    0.044    0.000    0.106    0.000 arraycolumn.py:40(size_unpack)
       38    0.032    0.001    0.032    0.001 {method 'read' of '_ssl._SSLSocket' objects}
    52513    0.025    0.000    0.025    0.000 base.py:101(<genexpr>)
    30216    0.019    0.000    0.019    0.000 {method 'unpack' of 'Struct' objects}
        6    0.016    0.003    0.054    0.009 base.py:94(_read_data)
    60441    0.014    0.000    0.014    0.000 {method 'append' of 'list' objects}
    30215    0.011    0.000    0.011    0.000 {built-in method builtins.isinstance}
        1    0.010    0.010    0.430    0.430 <string>:1(<module>)
    30206    0.008    0.000    0.008    0.000 {method 'appendleft' of 'collections.deque' objects}
    30206    0.007    0.000    0.007    0.000 {method 'pop' of 'collections.deque' objects}
```