@dourvaris 2020-09-13
Cleaned up some non relevant changes made by python code formatter. Ready for review.